@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license.php
 """
@@ -9,6 +8,7 @@ Example:
     with WsgiDavTestServer(opts):
         ... test methods
 """
+
 import multiprocessing
 import os
 import shutil
@@ -44,7 +44,7 @@ class Timing:
 
     def __exit__(self, exc_type, exc_value, traceback):
         elap = time.time() - self.start
-        msg = ["Timing {:<20} took {:>6.3f} sec".format(repr(self.name), elap)]
+        msg = [f"Timing {repr(self.name):<20} took {elap:>6.3f} sec"]
         if self.count:
             fmt = self.fmt or "{:0,.1f} bytes/sec"
             msg.append(fmt.format(float(self.count) / elap))
@@ -92,8 +92,6 @@ def run_wsgidav_server(with_auth, with_ssl, provider=None, **kwargs):
     if provider is None:
         provider = FilesystemProvider(share_path)
 
-    # config = DEFAULT_CONFIG.copy()
-    # config.update({
     config = {
         "host": "127.0.0.1",
         "port": 8080,
@@ -216,9 +214,7 @@ class WsgiDavTestServer:
         # time.sleep(self.start_delay)
         if not self.startup_event.wait(self.startup_timeout):
             raise RuntimeError(
-                "WsgiDavTestServer start() timed out after {} seconds".format(
-                    self.startup_timeout
-                )
+                f"WsgiDavTestServer start() timed out after {self.startup_timeout} seconds"
             )
         print("Starting WsgiDavTestServer... running.")
         return self
