@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Original PyFileServer (c) 2005 Ho Chun Wei.
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license.php
@@ -15,10 +14,12 @@ This provider creates instances of :class:`~wsgidav.fs_dav_provider.FileResource
 and :class:`~wsgidav.fs_dav_provider.FolderResource` to represent files and
 directories respectively.
 """
+
 import os
 import shutil
 import stat
 import sys
+from typing import List
 
 from wsgidav import util
 from wsgidav.dav_error import HTTP_FORBIDDEN, DAVError
@@ -40,12 +41,12 @@ class FileResource(DAVNonCollection):
     See also _DAVResource, DAVNonCollection, and FilesystemProvider.
     """
 
-    def __init__(self, path, environ, file_path):
+    def __init__(self, path: str, environ: dict, file_path: str):
         super().__init__(path, environ)
-        self._file_path = file_path
-        self.file_stat = os.stat(self._file_path)
+        self._file_path: str = file_path
+        self.file_stat: os.stat_result = os.stat(self._file_path)
         # Setting the name from the file path should fix the case on Windows
-        self.name = os.path.basename(self._file_path)
+        self.name: str = os.path.basename(self._file_path)
         self.name = util.to_str(self.name)
 
     # Getter methods for standard live properties
@@ -66,6 +67,9 @@ class FileResource(DAVNonCollection):
 
     def get_last_modified(self):
         return self.file_stat[stat.ST_MTIME]
+
+    def is_link(self):
+        return os.path.islink(self._file_path)
 
     def support_etag(self):
         return True
@@ -143,7 +147,7 @@ class FileResource(DAVNonCollection):
         fpDest = self.provider._loc_to_file_path(dest_path, self.environ)
         assert not util.is_equal_or_child_uri(self.path, dest_path)
         assert not os.path.exists(fpDest)
-        _logger.debug("move_recursive({}, {})".format(self._file_path, fpDest))
+        _logger.debug(f"move_recursive({self._file_path}, {fpDest})")
         shutil.move(self._file_path, fpDest)
         # (Live properties are copied by copy2 or copystat)
         # Move dead properties
@@ -174,10 +178,11 @@ class FolderResource(DAVCollection):
     See also _DAVResource, DAVCollection, and FilesystemProvider.
     """
 
-    def __init__(self, path, environ, file_path):
+    def __init__(self, path: str, environ: dict, file_path):
         super().__init__(path, environ)
-        self._file_path = file_path
-        self.file_stat = os.stat(self._file_path)
+        self._file_path: str = file_path
+        self.file_stat: os.stat_result = os.stat(self._file_path)
+        self.fs_opts = self.provider
         # Setting the name from the file path should fix the case on Windows
         self.name = os.path.basename(self._file_path)
         self.name = util.to_str(self.name)  # .encode("utf8")
@@ -195,10 +200,19 @@ class FolderResource(DAVCollection):
     def get_etag(self):
         return None
 
+    def get_used_bytes(self):
+        return shutil.disk_usage(self._file_path).used
+
+    def get_available_bytes(self):
+        return shutil.disk_usage(self._file_path).free
+
     def get_last_modified(self):
         return self.file_stat[stat.ST_MTIME]
 
-    def get_member_names(self):
+    def is_link(self):
+        return os.path.islink(self._file_path)
+
+    def get_member_names(self) -> List[str]:
         """Return list of direct collection member names (utf-8 encoded).
 
         See DAVCollection.get_member_names()
@@ -221,30 +235,36 @@ class FolderResource(DAVCollection):
             assert util.is_str(name)
             # Skip non files (links and mount points)
             fp = os.path.join(self._file_path, name)
+            # if not self.provider.fs_opts.get("follow_symlinks") and os.path.islink(fp):
+            #     _logger.info(f"Skipping symlink {fp!r}")
+            #     continue
             if not os.path.isdir(fp) and not os.path.isfile(fp):
-                _logger.debug("Skipping non-file {!r}".format(fp))
+                _logger.info(f"Skipping non-file {fp!r}")
                 continue
             # name = name.encode("utf8")
             name = util.to_str(name)
             nameList.append(name)
         return nameList
 
-    def get_member(self, name):
+    def get_member(self, name: str) -> FileResource:
         """Return direct collection member (DAVResource or derived).
 
         See DAVCollection.get_member()
         """
-        assert util.is_str(name), "{!r}".format(name)
+        assert util.is_str(name), f"{name!r}"
         fp = os.path.join(self._file_path, util.to_str(name))
         # name = name.encode("utf8")
         path = util.join_uri(self.path, name)
+        res = None
+        # if not self.provider.fs_opts.get("follow_symlinks") and os.path.islink(fp):
+        #     _logger.info(f"Skipping symlink {path}")
+        # elif
         if os.path.isdir(fp):
             res = FolderResource(path, self.environ, fp)
         elif os.path.isfile(fp):
             res = FileResource(path, self.environ, fp)
         else:
-            _logger.debug("Skipping non-file {}".format(path))
-            res = None
+            _logger.debug(f"Skipping non-file {path}")
         return res
 
     # --- Read / write -------------------------------------------------------
@@ -300,7 +320,7 @@ class FolderResource(DAVCollection):
             # u'C:\\temp\\litmus\\ccdest'
             shutil.copystat(self._file_path, fpDest)
         except Exception:
-            _logger.exception("Could not copy folder stats: {}".format(self._file_path))
+            _logger.exception(f"Could not copy folder stats: {self._file_path}")
         # (Live properties are copied by copy2 or copystat)
         # Copy dead properties
         propMan = self.provider.prop_manager
@@ -329,7 +349,7 @@ class FolderResource(DAVCollection):
         fpDest = self.provider._loc_to_file_path(dest_path, self.environ)
         assert not util.is_equal_or_child_uri(self.path, dest_path)
         assert not os.path.exists(fpDest)
-        _logger.debug("move_recursive({}, {})".format(self._file_path, fpDest))
+        _logger.debug(f"move_recursive({self._file_path}, {fpDest})")
         shutil.move(self._file_path, fpDest)
         # (Live properties are copied by copy2 or copystat)
         # Move dead properties
@@ -355,33 +375,42 @@ class FolderResource(DAVCollection):
 # FilesystemProvider
 # ========================================================================
 class FilesystemProvider(DAVProvider):
-    def __init__(self, root_folder, *, readonly=False, shadow=None):
-        # and resolve relative to config file
-        # root_folder = os.path.expandvars(os.xpath.expanduser(root_folder))
+    """Default implementation of a filesystem DAVProvider.
+
+    Args:
+        root_folder (str)
+        readonly (bool)
+        fs_opts (dict | None): defaults to `config.fs_dav_provider`
+        shadow (dict):  @deprecated (Use option `fs_provider.shadow_map` instead)
+    """
+
+    def __init__(self, root_folder, *, readonly=False, fs_opts=None):
+        # root_folder is typically already resolved relative to config file
+        # and has user ~ expanded
         root_folder = os.path.abspath(root_folder)
         if not root_folder or not os.path.exists(root_folder):
-            raise ValueError("Invalid root path: {}".format(root_folder))
+            raise ValueError(f"Invalid root path: {root_folder}")
 
         super().__init__()
 
         self.root_folder_path = root_folder
         self.readonly = readonly
-        if shadow:
-            self.shadow = {k.lower(): v for k, v in shadow.items()}
-        else:
-            self.shadow = {}
+        if fs_opts is None:
+            _logger.warning(f"{self}: no `fs_opts` parameter passed to constructor.")
+            fs_opts = {}
+        self.fs_opts = fs_opts
+        # Get shadow map and convert keys to lower case
+        self.shadow_map = self.fs_opts.get("shadow_map") or {}
+        if self.shadow_map:
+            self.shadow_map = {k.lower(): v for k, v in self.shadow_map.items()}
 
     def __repr__(self):
-        rw = "Read-Write"
-        if self.readonly:
-            rw = "Read-Only"
-        return "{} for path '{}' ({})".format(
-            self.__class__.__name__, self.root_folder_path, rw
-        )
+        rw = "Read-Only" if self.readonly else "Read-Write"
+        return f"{self.__class__.__name__} for path {self.root_folder_path!r} ({rw})"
 
-    def _resolve_shadow_path(self, path, environ, file_path):
+    def _resolve_shadow_path(self, path: str, environ: dict, file_path):
         """File not found: See if there is a shadow configured."""
-        shadow = self.shadow.get(path.lower())
+        shadow = self.shadow_map.get(path.lower())
         # _logger.info(f"Shadow {path} -> {shadow} {self.shadow}")
         if not shadow:
             return False, file_path
@@ -401,7 +430,7 @@ class FilesystemProvider(DAVProvider):
         _logger.info(f"Shadow {path} -> {shadow}")
         return True, shadow
 
-    def _loc_to_file_path(self, path, environ=None):
+    def _loc_to_file_path(self, path: str, environ: dict = None):
         """Convert resource path to a unicode absolute file path.
         Optional environ argument may be useful e.g. in relation to per-user
         sub-folder chrooting inside root_folder_path.
@@ -413,31 +442,31 @@ class FilesystemProvider(DAVProvider):
 
         path_parts = path.strip("/").split("/")
         file_path = os.path.abspath(os.path.join(root_path, *path_parts))
+
+        # Try alternative URL if not found (or even override target):
         is_shadow, file_path = self._resolve_shadow_path(path, environ, file_path)
+
         if not file_path.startswith(root_path) and not is_shadow:
             raise RuntimeError(
-                "Security exception: tried to access file outside root: {}".format(
-                    file_path
-                )
+                f"Security exception: tried to access file outside root: {file_path}"
             )
 
         # Convert to unicode
         file_path = util.to_unicode_safe(file_path)
         return file_path
 
-    def is_readonly(self):
-        return self.readonly
-
-    def get_resource_inst(self, path, environ):
+    def get_resource_inst(self, path: str, environ: dict) -> FileResource:
         """Return info dictionary for path.
 
         See DAVProvider.get_resource_inst()
         """
         self._count_get_resource_inst += 1
         fp = self._loc_to_file_path(path, environ)
+
         if not os.path.exists(fp):
             return None
-
+        if not self.fs_opts.get("follow_symlinks") and os.path.islink(fp):
+            raise DAVError(HTTP_FORBIDDEN, f"Symlink support is disabled: {fp!r}")
         if os.path.isdir(fp):
             return FolderResource(path, environ, fp)
         return FileResource(path, environ, fp)
