@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Licensed under the MIT license: http://www.opensource.org/licenses/mit-license.php
 """
 Implements a property manager based on CouchDB.
@@ -22,6 +21,7 @@ Valid options are (sample shows defaults)::
             }
 
 """
+
 from urllib.parse import quote
 from uuid import uuid4
 
@@ -116,7 +116,7 @@ class CouchPropertyManager:
         """Return properties document for path."""
         # Query the permanent view to find a url
         vr = self.db.view("properties/by_url", key=url, include_docs=True)
-        _logger.debug("find(%r) returned %s" % (url, len(vr)))
+        _logger.debug(f"find({url!r}) returned {len(vr)}")
         assert len(vr) <= 1, "Found multiple matches for %r" % url
         for row in vr:
             assert row.doc
@@ -131,9 +131,7 @@ class CouchPropertyManager:
                 if(doc.type === 'properties' && url.indexOf('%s') === 0) {
                     emit(doc.url, { 'id': doc._id, 'url': doc.url });
                 }
-            }""" % (
-            url + "/"
-        )
+            }""" % (url + "/")
         vr = self.db.query(map_fun, include_docs=True)
         for row in vr:
             yield row.doc
@@ -149,7 +147,7 @@ class CouchPropertyManager:
         return propNames
 
     def get_property(self, norm_url, name, environ=None):
-        _logger.debug("get_property(%s, %s)" % (norm_url, name))
+        _logger.debug(f"get_property({norm_url}, {name})")
         doc = self._find(norm_url)
         if not doc:
             return None
@@ -184,7 +182,7 @@ class CouchPropertyManager:
         self.db.save(doc)
 
     def remove_property(self, norm_url, name, dry_run=False, environ=None):
-        _logger.debug("remove_property(%s, %s, dry_run=%s)" % (norm_url, name, dry_run))
+        _logger.debug(f"remove_property({norm_url}, {name}, dry_run={dry_run})")
         if dry_run:
             # TODO: can we check anything here?
             return
@@ -206,10 +204,10 @@ class CouchPropertyManager:
         doc = self._find(srcUrl)
         if not doc:
             _logger.debug(
-                "copy_properties(%s, %s): src has no properties" % (srcUrl, destUrl)
+                f"copy_properties({srcUrl}, {destUrl}): src has no properties"
             )
             return
-        _logger.debug("copy_properties(%s, %s)" % (srcUrl, destUrl))
+        _logger.debug(f"copy_properties({srcUrl}, {destUrl})")
         assert not self._find(destUrl)
         doc2 = {
             "_id": uuid4().hex,
@@ -221,13 +219,13 @@ class CouchPropertyManager:
         self.db.save(doc2)
 
     def move_properties(self, srcUrl, destUrl, with_children, environ=None):
-        _logger.debug("move_properties(%s, %s, %s)" % (srcUrl, destUrl, with_children))
+        _logger.debug(f"move_properties({srcUrl}, {destUrl}, {with_children})")
         if with_children:
             # Match URLs that are equal to <srcUrl> or begin with '<srcUrl>/'
             docList = self._find_descendents(srcUrl)
             for doc in docList:
                 newDest = doc["url"].replace(srcUrl, destUrl)
-                _logger.debug("move property %s -> %s" % (doc["url"], newDest))
+                _logger.debug("move property {} -> {}".format(doc["url"], newDest))
                 doc["url"] = newDest
                 self.db.save(doc)
         else:
@@ -235,7 +233,7 @@ class CouchPropertyManager:
             # TODO: use findAndModify()?
             doc = self._find(srcUrl)
             if doc:
-                _logger.debug("move property %s -> %s" % (doc["url"], destUrl))
+                _logger.debug("move property {} -> {}".format(doc["url"], destUrl))
                 doc["url"] = destUrl
                 self.db.save(doc)
         return
