@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Licensed under the MIT license: http://www.opensource.org/licenses/mit-license.php
 """
 Implements a property manager based on MongoDB.
@@ -22,6 +21,7 @@ Valid options are (sample shows defaults)::
             }
 
 """
+
 from urllib.parse import quote
 
 import pymongo
@@ -36,7 +36,7 @@ _logger = util.get_module_logger(__name__)
 HIDDEN_KEYS = ("_id", "_url", "_title")
 
 # MongiDB doesn't accept '.' in key names, so we have to escape it.
-# Use a key that is unlikely to occur in proprty names
+# Use a key that is unlikely to occur in property names
 DOT_ESCAPE = "^"
 
 
@@ -113,7 +113,7 @@ class MongoPropertyManager:
         return propNames
 
     def get_property(self, norm_url, name, environ=None):
-        _logger.debug("get_property(%s, %s)" % (norm_url, name))
+        _logger.debug(f"get_property({norm_url}, {name})")
         doc = self.collection.find_one({"_url": norm_url})
         if not doc:
             return None
@@ -143,7 +143,7 @@ class MongoPropertyManager:
 
     def remove_property(self, norm_url, name, dry_run=False, environ=None):
         """ """
-        _logger.debug("remove_property(%s, %s, dry_run=%s)" % (norm_url, name, dry_run))
+        _logger.debug(f"remove_property({norm_url}, {name}, dry_run={dry_run})")
         if dry_run:
             # TODO: can we check anything here?
             return
@@ -165,15 +165,15 @@ class MongoPropertyManager:
         doc = self.collection.find_one({"_url": srcUrl})
         if not doc:
             _logger.debug(
-                "copy_properties(%s, %s): src has no properties" % (srcUrl, destUrl)
+                f"copy_properties({srcUrl}, {destUrl}): src has no properties"
             )
             return
-        _logger.debug("copy_properties(%s, %s)" % (srcUrl, destUrl))
+        _logger.debug(f"copy_properties({srcUrl}, {destUrl})")
         doc2 = doc.copy()
         self.collection.insert(doc2)
 
     def move_properties(self, srcUrl, destUrl, with_children, environ=None):
-        _logger.debug("move_properties(%s, %s, %s)" % (srcUrl, destUrl, with_children))
+        _logger.debug(f"move_properties({srcUrl}, {destUrl}, {with_children})")
         if with_children:
             # Match URLs that are equal to <srcUrl> or begin with '<srcUrl>/'
             matchBegin = "^" + srcUrl.rstrip("/") + "/"
@@ -181,7 +181,7 @@ class MongoPropertyManager:
             docList = self.collection.find(query)
             for doc in docList:
                 newDest = doc["_url"].replace(srcUrl, destUrl)
-                _logger.debug("move property %s -> %s" % (doc["_url"], newDest))
+                _logger.debug("move property {} -> {}".format(doc["_url"], newDest))
                 doc["_url"] = newDest
                 self.collection.save(doc)
         else:
@@ -189,7 +189,7 @@ class MongoPropertyManager:
             # TODO: use findAndModify()?
             doc = self.collection.find_one({"_url": srcUrl})
             if doc:
-                _logger.debug("move property %s -> %s" % (doc["_url"], destUrl))
+                _logger.debug("move property {} -> {}".format(doc["_url"], destUrl))
                 doc["_url"] = destUrl
                 self.collection.save(doc)
         return
