@@ -1,9 +1,9 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Licensed under the MIT license: http://www.opensource.org/licenses/mit-license.php
 """
 Tools that make it easier to implement custom WsgiDAV providers.
 """
+
 import os
 import stat
 
@@ -141,7 +141,7 @@ class FileResource(_VirtualNonCollection):
 
     def __init__(self, path, environ, file_path):
         if not os.path.exists(file_path):
-            _logger.error("FileResource({!r}) does not exist.".format(file_path))
+            _logger.error(f"FileResource({file_path!r}) does not exist.")
         super().__init__(path, environ)
         self.file_path = file_path
 
