@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Original PyFileServer (c) 2005 Ho Chun Wei.
 # Licensed under the MIT license: http://www.opensource.org/licenses/mit-license.php
 """
@@ -57,6 +56,7 @@ its limitations:
       unnecessary queries to the database.
 
 """
+
 import csv
 import hashlib
 import time
@@ -259,7 +259,7 @@ class MySQLBrowserResource(_DAVResource):
             conn = self.provider._init_connection()
             fieldlist = self.provider._get_field_list(conn, tableName)
             for fieldname in fieldlist:
-                propNames.append("{%s:}%s" % (tableName, fieldname))
+                propNames.append(f"{{{tableName}:}}{fieldname}")
             conn.close()
         return propNames
 
@@ -317,12 +317,7 @@ class MySQLBrowserProvider(DAVProvider):
         self._count_initConnection = 0
 
     def __repr__(self):
-        return "%s for db '%s' on '%s' (user: '%s')'" % (
-            self.__class__.__name__,
-            self._db,
-            self._host,
-            self._user,
-        )
+        return f"{self.__class__.__name__} for db '{self._db}' on '{self._host}' (user: '{self._user}')'"
 
     def _split_path(self, path):
         """Return (tableName, primaryKey) tuple for a request path."""
