@@ -1,11 +1,11 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/m ar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/m ar10/wsgidav
 # Original PyFileServer (c) 2005 Ho Chun Wei.
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license.php
 """
 Miscellaneous support functions for WsgiDAV.
 """
+
 import base64
 import calendar
 import collections.abc
@@ -13,17 +13,18 @@ import logging
 import mimetypes
 import os
 import re
-import socket
 import stat
 import sys
 import time
+import warnings
 from copy import deepcopy
 from email.utils import formatdate, parsedate
 from hashlib import md5
 from pprint import pformat
-from typing import Iterable, Optional
+from typing import Iterable, Optional, Tuple
 from urllib.parse import quote
 
+from wsgidav import __version__
 from wsgidav.dav_error import (
     HTTP_BAD_REQUEST,
     HTTP_CREATED,
@@ -31,6 +32,7 @@ from wsgidav.dav_error import (
     HTTP_NOT_MODIFIED,
     HTTP_OK,
     HTTP_PRECONDITION_FAILED,
+    HTTP_RANGE_NOT_SATISFIABLE,
     DAVError,
     as_DAVError,
     get_http_status_string,
@@ -43,15 +45,36 @@ __docformat__ = "reStructuredText"
 BASE_LOGGER_NAME = "wsgidav"
 _logger = logging.getLogger(BASE_LOGGER_NAME)
 
-PYTHON_VERSION = "{}.{}.{}".format(
-    sys.version_info[0], sys.version_info[1], sys.version_info[2]
-)
+#: Currently used Python version as string
+PYTHON_VERSION = ".".join([str(s) for s in sys.version_info[:3]])
 
 filesystemencoding = sys.getfilesystemencoding()
+
+#: Project name and version presented to the clients
+#: This is reset to ``"WsgiDAV"`` if ``suppress_version_info`` is set in the
+#: configuration.
+public_wsgidav_info = f"WsgiDAV/{__version__}"
+#: This is reset to ``"Python/3"`` if ``suppress_version_info`` is set in the
+#: configuration.
+public_python_info = f"Python/{PYTHON_VERSION}"
 
 
 class NO_DEFAULT:
     """"""
+
+
+def check_python_version(min_version: Tuple[str]) -> bool:
+    """Check for deprecated Python version."""
+    if sys.version_info < min_version:
+        min_ver = ".".join([str(s) for s in min_version[:3]])
+        warnings.warn(
+            f"Support for Python version less than `{min_ver}` is deprecated "
+            f"(using {PYTHON_VERSION})",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return False
+    return True
 
 
 # ========================================================================
@@ -157,11 +180,11 @@ def get_dict_value(d, key_path, default=NO_DEFAULT, *, as_dict=False):
             seg = seg[1:-1]
             value = value[int(seg)]
         else:
-            # raise ValueError("Segment '{}' cannot be nested".format(seg))
+            # raise ValueError("Segment {!r} cannot be nested".format(seg))
             try:
                 value = getattr(value, seg)
             except AttributeError:
-                raise  # ValueError("Segment '{}' cannot be nested".format(seg))
+                raise  # ValueError("Segment {!r} cannot be nested".format(seg))
 
     return value
 
@@ -214,12 +237,12 @@ def check_tags(tags, known, *, msg=None, raise_error=True, required=False):
     res = []
     unknown = tags.difference(known)
     if unknown:
-        res.append("Unknown: '{}'".format("', '".join(unknown)))
+        res.append("Unknown: {!r}".format("', '".join(unknown)))
 
     if required:
         missing = required.difference(tags)
         if missing:
-            res.append("Missing: '{}'".format("', '".join(missing)))
+            res.append("Missing: {!r}".format("', '".join(missing)))
 
     if res:
         if msg:
@@ -227,14 +250,14 @@ def check_tags(tags, known, *, msg=None, raise_error=True, required=False):
 
         if required and optional:
             res.append(
-                "Required: ('{}'). Optional: ('{}')".format(
+                "Required: ({!r}). Optional: ({!r})".format(
                     "', '".join(required), "', '".join(optional)
                 )
             )
         elif required:
-            res.append("Required: ('{}')".format("', '".join(required)))
+            res.append("Required: ({!r})".format("', '".join(required)))
         elif optional:
-            res.append("Optional: ('{}')".format("', '".join(optional)))
+            res.append("Optional: ({!r})".format("', '".join(optional)))
 
         res = "\n".join(res)
         if raise_error:
@@ -366,11 +389,13 @@ def init_logging(config):
     The base logger is filtered by the `verbose` configuration option.
     Log entries will have a time stamp and thread id.
 
+    **Note:** init_logging() is automatically called if an application adds
+    ``"logging": { "enable": true }`` to the configuration.
 
     Module loggers
     ~~~~~~~~~~~~~~
-    Module loggers (e.g 'wsgidav.lock_man.lock_manager') are named loggers, that can be
-    independently switched to DEBUG mode.
+    Module loggers (e.g 'wsgidav.lock_man.lock_manager') are named loggers, that
+    can be independently switched to DEBUG mode.
 
     Except for verbosity, they will inherit settings from the base logger.
 
@@ -384,7 +409,7 @@ def init_logging(config):
 
         _logger = util.get_module_logger(__name__)
         [..]
-        _logger.debug("foo: '{}'".format(s))
+        _logger.debug("foo: {!r}".format(s))
 
     This logger would be enabled by passing its name to init_logging()::
 
@@ -529,7 +554,7 @@ def dynamic_import_class(name):
     try:
         module = importlib.import_module(module_name)
     except Exception as e:
-        _logger.error("Dynamic import of {!r} failed: {}".format(name, e))
+        _logger.error(f"Dynamic import of {name!r} failed: {e}")
         raise
     the_class = getattr(module, class_name)
     return the_class
@@ -710,9 +735,7 @@ def to_unicode_safe(s):
     try:
         u = to_str(s, "utf8")
     except ValueError:
-        _logger.error(
-            "to_unicode_safe({!r}) *** UTF-8 failed. Trying ISO-8859-1".format(s)
-        )
+        _logger.error(f"to_unicode_safe({s!r}) *** UTF-8 failed. Trying ISO-8859-1")
         u = to_str(s, "ISO-8859-1")
     return u
 
@@ -739,13 +762,13 @@ def safe_re_encode(s, encoding_to, *, errors="backslashreplace"):
 def string_repr(s):
     """Return a string as hex dump."""
     if is_bytes(s):
-        res = "{!r}: ".format(s)
+        res = f"{s!r}: "
         for b in s:
             if type(b) is str:  # Py2
                 b = ord(b)
             res += "%02x " % b
         return res
-    return "{}".format(s)
+    return f"{s}"
 
 
 def get_file_extension(path):
@@ -785,11 +808,11 @@ def byte_number_string(
         # locale.setlocale(locale.LC_ALL, "")
         # # TODO: make precision configurable
         # snum = locale.format("%d", number, thousandsSep)
-        snum = "{:,d}".format(number)
+        snum = f"{number:,d}"
     else:
         snum = str(number)
 
-    return "{}{}{}".format(snum, magsuffix, bytesuffix)
+    return f"{snum}{magsuffix}{bytesuffix}"
 
 
 def fix_path(path, root, *, expand_vars=True, must_exist=True, allow_none=True):
@@ -798,7 +821,6 @@ def fix_path(path, root, *, expand_vars=True, must_exist=True, allow_none=True):
     Convert path to absolute if required, expand leading '~' as user home dir,
     expand %VAR%, $Var, ...
     """
-    org_path = path
     if path in (None, ""):
         if allow_none:
             return None
@@ -814,6 +836,9 @@ def fix_path(path, root, *, expand_vars=True, must_exist=True, allow_none=True):
                 root = os.path.dirname(config_file)
             else:
                 root = os.getcwd()
+        # NOTE:
+        # Changed in version 3.13: On Windows, `os.path.isabs` returns False
+        # if the given path starts with exactly one (back)slash.
         path = os.path.abspath(os.path.join(root, path))
 
     if expand_vars:
@@ -822,8 +847,6 @@ def fix_path(path, root, *, expand_vars=True, must_exist=True, allow_none=True):
     if must_exist and not os.path.exists(path):
         raise ValueError(f"Invalid path: {path!r}")
 
-    if org_path != path:
-        print(f"fix_path({org_path}) => {path}")
     return path
 
 
@@ -895,9 +918,7 @@ def read_and_discard_input(environ):
                 n = 1
             body = wsgi_input.read(n)
             _logger.debug(
-                "Reading {} bytes from potentially unread httpserver.LimitedLengthFile: '{}'...".format(
-                    n, body[:50]
-                )
+                f"Reading {n} bytes from potentially unread httpserver.LimitedLengthFile: {body[:50]!r}..."
             )
 
     elif hasattr(wsgi_input, "_sock") and hasattr(wsgi_input._sock, "settimeout"):
@@ -915,20 +936,25 @@ def read_and_discard_input(environ):
                     n = 1
                 body = wsgi_input.read(n)
                 _logger.debug(
-                    "Reading {} bytes from potentially unread POST body: '{}'...".format(
-                        n, body[:50]
-                    )
+                    f"Reading {n} bytes from potentially unread POST body: {body[:50]!r}..."
                 )
-            except socket.error as se:
+            except OSError as se:
                 # se(10035, 'The socket operation could not complete without blocking')
-                _logger.error("-> read {} bytes failed: {}".format(n, se))
+                _logger.error(f"-> read {n} bytes failed: {se}")
             # Restore socket settings
             sock.settimeout(timeout)
         except Exception:
-            _logger.error("--> wsgi_input.read(): {}".format(sys.exc_info()))
+            _logger.error(f"--> wsgi_input.read(): {sys.exc_info()}")
 
 
-def fail(value, context_info=None, *, src_exception=None, err_condition=None):
+def fail(
+    value,
+    context_info=None,
+    *,
+    src_exception=None,
+    err_condition=None,
+    add_headers=None,
+):
     """Wrapper to raise (and log) DAVError."""
     if isinstance(value, Exception):
         e = as_DAVError(value)
@@ -938,8 +964,9 @@ def fail(value, context_info=None, *, src_exception=None, err_condition=None):
             context_info,
             src_exception=src_exception,
             err_condition=err_condition,
+            add_headers=add_headers,
         )
-    _logger.debug("Raising DAVError {}".format(e.get_user_info()))
+    _logger.debug(f"Raising DAVError {e.get_user_info()}")
     raise e
 
 
@@ -1131,7 +1158,7 @@ def parse_xml_body(environ, *, allow_empty=False):
             if content_length < 0:
                 raise DAVError(HTTP_BAD_REQUEST, "Negative content-length.")
         except ValueError:
-            raise DAVError(HTTP_BAD_REQUEST, "content-length is not numeric.")
+            raise DAVError(HTTP_BAD_REQUEST, "content-length is not numeric.") from None
 
         if content_length == 0:
             requestbody = ""
@@ -1148,7 +1175,9 @@ def parse_xml_body(environ, *, allow_empty=False):
     try:
         rootEL = etree.fromstring(requestbody)
     except Exception as e:
-        raise DAVError(HTTP_BAD_REQUEST, "Invalid XML format.", src_exception=e)
+        raise DAVError(
+            HTTP_BAD_REQUEST, "Invalid XML format.", src_exception=e
+        ) from None
 
     # If dumps of the body are desired, then this is the place to do it pretty:
     if environ.get("wsgidav.dump_request_body"):
@@ -1196,6 +1225,8 @@ def send_status_response(
     headers = []
     if add_headers:
         headers.extend(add_headers)
+    if isinstance(e, DAVError) and e.add_headers:
+        headers.extend(e.add_headers)
     #    if 'keep-alive' in environ.get('HTTP_CONNECTION', '').lower():
     #        headers += [
     #            ('Connection', 'keep-alive'),
@@ -1292,7 +1323,7 @@ def add_property_response(multistatus_elem, href, prop_list):
         ns, _ = split_namespace(name)
         if ns != "DAV:" and ns not in nsDict and ns != "":
             nsDict[ns] = True
-            nsMap["NS{}".format(nsCount)] = ns
+            nsMap[f"NS{nsCount}"] = ns
             nsCount += 1
 
         propDict.setdefault(status, []).append((name, value))
@@ -1322,7 +1353,7 @@ def add_property_response(multistatus_elem, href, prop_list):
                 #                etree.SubElement(propEL, name).text = value
                 etree.SubElement(propEL, name).text = to_unicode_safe(value)
         # <status>
-        etree.SubElement(propstatEL, "{DAV:}status").text = "HTTP/1.1 {}".format(status)
+        etree.SubElement(propstatEL, "{DAV:}status").text = f"HTTP/1.1 {status}"
 
 
 # ========================================================================
@@ -1407,17 +1438,11 @@ def get_file_etag(file_path):
 
     fstat = os.stat(unicode_file_path)
     if sys.platform == "win32":
-        etag = "{}-{}-{}".format(
-            md5(file_path).hexdigest(),
-            fstat[stat.ST_MTIME],
-            fstat[stat.ST_SIZE],
+        etag = (
+            f"{md5(file_path).hexdigest()}-{fstat[stat.ST_MTIME]}-{fstat[stat.ST_SIZE]}"
         )
     else:
-        etag = "{}-{}-{}".format(
-            fstat[stat.ST_INO],
-            fstat[stat.ST_MTIME],
-            fstat[stat.ST_SIZE],
-        )
+        etag = f"{fstat[stat.ST_INO]}-{fstat[stat.ST_MTIME]}-{fstat[stat.ST_SIZE]}"
     return etag
 
 
@@ -1428,68 +1453,80 @@ def get_file_etag(file_path):
 # Range Specifiers
 reByteRangeSpecifier = re.compile("(([0-9]+)-([0-9]*))")
 reSuffixByteRangeSpecifier = re.compile("(-([0-9]+))")
-# reByteRangeSpecifier = re.compile("(([0-9]+)\-([0-9]*))")
-# reSuffixByteRangeSpecifier = re.compile("(\-([0-9]+))")
 
 
-def obtain_content_ranges(rangetext, filesize):
+def obtain_content_ranges(range_header, filesize):
     """
-    returns tuple (list, value)
+    See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Range
 
-    list
+    Return tuple (range_list, total_length)
+
+    range_list
         content ranges as values to their parsed components in the tuple
         (seek_position/abs position of first byte, abs position of last byte, num_of_bytes_to_read)
-    value
+    total_length
         total length for Content-Length
     """
-    listReturn = []
-    seqRanges = rangetext.split(",")
-    for subrange in seqRanges:
-        matched = False
-        if not matched:
-            mObj = reByteRangeSpecifier.search(subrange)
-            if mObj:
-                firstpos = int(mObj.group(2))
-                if mObj.group(3) == "":
-                    lastpos = filesize - 1
-                else:
-                    lastpos = int(mObj.group(3))
-                if firstpos <= lastpos and firstpos < filesize:
-                    if lastpos >= filesize:
-                        lastpos = filesize - 1
-                    listReturn.append((firstpos, lastpos))
-                    matched = True
-        if not matched:
-            mObj = reSuffixByteRangeSpecifier.search(subrange)
-            if mObj:
-                firstpos = filesize - int(mObj.group(2))
-                if firstpos < 0:
-                    firstpos = 0
-                lastpos = filesize - 1
-                listReturn.append((firstpos, lastpos))
+    list_ranges = []
+    request_ranges = range_header.split(",")
+    for subrange in request_ranges:
+        is_matched = False
+        if not is_matched:
+            match = reByteRangeSpecifier.search(subrange)
+            if match:
+                range_start = int(match.group(2))
 
-                matched = True
+                if range_start >= filesize:
+                    # TODO: set "Content-Range: bytes */filesize"
+                    fail(
+                        HTTP_RANGE_NOT_SATISFIABLE,
+                        f"Requested range starts behind file size ({range_start} >= {filesize})",
+                        add_headers=[("Content-Range", f"bytes */{filesize}")],
+                    )
+
+                if match.group(3) == "":
+                    # "START-"
+                    range_end = filesize - 1
+                else:
+                    # "START-END"
+                    range_end = int(match.group(3))
+
+                if range_start <= range_end and range_start < filesize:
+                    if range_end >= filesize:
+                        range_end = filesize - 1
+                    list_ranges.append((range_start, range_end))
+                    is_matched = True
+
+        if not is_matched:
+            match = reSuffixByteRangeSpecifier.search(subrange)
+            if match:
+                range_start = filesize - int(match.group(2))
+                if range_start < 0:
+                    range_start = 0
+                range_end = filesize - 1
+                list_ranges.append((range_start, range_end))
+                is_matched = True
 
     # consolidate ranges
-    listReturn.sort()
-    listReturn2 = []
-    totallength = 0
-    while len(listReturn) > 0:
-        (rfirstpos, rlastpos) = listReturn.pop()
-        counter = len(listReturn)
+    list_ranges.sort()
+    list_ranges_2 = []
+    total_length = 0
+    while len(list_ranges) > 0:
+        (rfirstpos, rlastpos) = list_ranges.pop()
+        counter = len(list_ranges)
         while counter > 0:
-            (nfirstpos, nlastpos) = listReturn[counter - 1]
+            (nfirstpos, nlastpos) = list_ranges[counter - 1]
             if nlastpos < rfirstpos - 1 or nfirstpos > nlastpos + 1:
                 pass
             else:
                 rfirstpos = min(rfirstpos, nfirstpos)
                 rlastpos = max(rlastpos, nlastpos)
-                del listReturn[counter - 1]
+                del list_ranges[counter - 1]
             counter = counter - 1
-        listReturn2.append((rfirstpos, rlastpos, rlastpos - rfirstpos + 1))
-        totallength = totallength + rlastpos - rfirstpos + 1
+        list_ranges_2.append((rfirstpos, rlastpos, rlastpos - rfirstpos + 1))
+        total_length = total_length + rlastpos - rfirstpos + 1
 
-    return (listReturn2, totallength)
+    return (list_ranges_2, total_length)
 
 
 # ========================================================================
@@ -1525,7 +1562,7 @@ def read_timeout_value_header(timeoutvalue):
 # ========================================================================
 
 
-def evaluate_http_conditionals(dav_res, last_modified, entitytag, environ):
+def evaluate_http_conditionals(dav_res, last_modified, entity_tag, environ):
     """Handle 'If-...:' headers (but not 'If:' header).
 
     If-Match
@@ -1563,7 +1600,7 @@ def evaluate_http_conditionals(dav_res, last_modified, entitytag, environ):
         token_list = parse_if_match_header(environ["HTTP_IF_MATCH"])
         match = False
         for token in token_list:
-            if token == entitytag or token == "*":
+            if token == entity_tag or token == "*":
                 match = True
                 break
         if not match:
@@ -1585,7 +1622,7 @@ def evaluate_http_conditionals(dav_res, last_modified, entitytag, environ):
     if "HTTP_IF_NONE_MATCH" in environ and dav_res.support_etag():
         token_list = parse_if_match_header(environ["HTTP_IF_NONE_MATCH"])
         for token in token_list:
-            if token == entitytag or token == "*":
+            if token == entity_tag or token == "*":
                 # ETag matched. If it's a GET request and we don't have an
                 # conflicting If-Modified header, we return NOT_MODIFIED
                 if (
@@ -1638,9 +1675,7 @@ def parse_if_header_dict(environ):
     ifLockList = []
 
     resource1 = "*"
-    for (tmpURLVar, URLVar, _tmpContentVar, contentVar) in reIfSeparator.findall(
-        iftext
-    ):
+    for tmpURLVar, URLVar, _tmpContentVar, contentVar in reIfSeparator.findall(iftext):
         if tmpURLVar != "":
             resource1 = URLVar
         else:
@@ -1668,19 +1703,17 @@ def parse_if_header_dict(environ):
 
     environ["wsgidav.conditions.if"] = ifDict
     environ["wsgidav.ifLockTokenList"] = ifLockList
-    _logger.debug("parse_if_header_dict\n{}".format(pformat(ifDict)))
+    _logger.debug(f"parse_if_header_dict\n{pformat(ifDict)}")
     return
 
 
-def test_if_header_dict(dav_res, dictIf, fullurl, locktokenlist, entitytag):
-    _logger.debug(
-        "test_if_header_dict({}, {}, {})".format(fullurl, locktokenlist, entitytag)
-    )
+def test_if_header_dict(dav_res, if_dict, fullurl, locktoken_list, entity_tag):
+    _logger.debug(f"test_if_header_dict({fullurl}, {locktoken_list}, {entity_tag})")
 
-    if fullurl in dictIf:
-        listTest = dictIf[fullurl]
-    elif "*" in dictIf:
-        listTest = dictIf["*"]
+    if fullurl in if_dict:
+        listTest = if_dict[fullurl]
+    elif "*" in if_dict:
+        listTest = if_dict["*"]
     else:
         return True
 
@@ -1689,13 +1722,13 @@ def test_if_header_dict(dav_res, dictIf, fullurl, locktokenlist, entitytag):
     for listTestConds in listTest:
         matchfailed = False
 
-        for (testflag, checkstyle, checkvalue) in listTestConds:
+        for testflag, checkstyle, checkvalue in listTestConds:
             if checkstyle == "entity" and supportEntityTag:
-                testresult = entitytag == checkvalue
+                testresult = entity_tag == checkvalue
             elif checkstyle == "entity":
                 testresult = testflag
             elif checkstyle == "locktoken":
-                testresult = checkvalue in locktokenlist
+                testresult = checkvalue in locktoken_list
             else:  # unknown
                 testresult = True
             checkresult = testresult == testflag
@@ -1734,7 +1767,7 @@ def guess_mime_type(url):
     if not mimetype:
         ext = os.path.splitext(url)[1]
         mimetype = _MIME_TYPES.get(ext)
-        _logger.debug("mimetype({}): {}".format(url, mimetype))
+        _logger.debug(f"mimetype({url}): {mimetype}")
     if not mimetype:
         mimetype = "application/octet-stream"
     return mimetype
