@@ -1,9 +1,7 @@
-# -*- coding: utf-8 -*-
+# import os
+# import sys
 
-import os
-import sys
+# import pkg_resources
 
-import pkg_resources
-
-sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
-pkg_resources.require("wsgidav")
+# sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+# pkg_resources.require("wsgidav")
