@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 Current WsgiDAV version number.
 
@@ -15,7 +14,8 @@ NOTE:
     When pywin32 is installed, number must be a.b.c for MSI builds?
     "3.0.0a4" seems not to work in this case!
 """
-__version__ = "4.0.3-a1"
+
+__version__ = "4.3.4-a1"
 
 # make version accessible as 'wsgidav.__version__'
 # from wsgidav._version import __version__  # noqa: F401
