@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license.php
 """
@@ -7,6 +6,7 @@ WSGI middleware used for CORS support (optional).
 
 Respond to CORS preflight OPTIONS request and inject CORS headers.
 """
+
 from wsgidav import util
 from wsgidav.mw.base_mw import BaseMiddleware
 
