@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# (c) 2009-2022 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
+# (c) 2009-2024 Martin Wendt and contributors; see WsgiDAV https://github.com/mar10/wsgidav
 # Licensed under the MIT license: http://www.opensource.org/licenses/mit-license.php
 """
 Implement the FileLikeQueue helper class.
@@ -16,6 +15,7 @@ consumer at the same time::
         return queue
 
 """
+
 import queue
 
 from wsgidav import util
@@ -67,7 +67,7 @@ class FileLikeQueue:
         # Get next chunk, cumulating requested size as needed
         while res == b"" or size < 0 or (size > 0 and len(res) < size):
             try:
-                # Read pending data, blocking if neccessary
+                # Read pending data, blocking if necessary
                 # (but handle the case that close() is called while waiting)
                 res += self.queue.get(True, 0.1)
             except queue.Empty:
@@ -128,7 +128,7 @@ class StreamingFile:
     """A file object wrapped around an iterator / data stream."""
 
     def __init__(self, data_stream):
-        """Intialise the object with the data stream."""
+        """Initialise the object with the data stream."""
         self.data_stream = data_stream
         self.buffer = ""
 
